Split value formatters into their own file
