+++ import/chrome.ts
@@ -1,5 +1,6 @@
-import {Profile, TimeFormatter, FrameInfo, CallTreeProfileBuilder} from '../profile'
+import {Profile, FrameInfo, CallTreeProfileBuilder} from '../profile'
 import {getOrInsert, lastOf} from '../utils'
+import {TimeFormatter} from '../value-formatters'
 
 interface TimelineEvent {
   pid: number
