+++ profile.ts
@@ -1,4 +1,5 @@
 import {lastOf, getOrInsert} from './utils'
+import {ValueFormatter, RawValueFormatter} from './value-formatters'
 const demangleCppModule = import('./demangle-cpp')
 
 // Force eager loading of the module
@@ -82,49 +83,6 @@ const rootFrame = new Frame({
   key: '(speedscope root)',
   name: '(speedscope root)',
 })
-
-export interface ValueFormatter {
-  format(v: number): string
-}
-
-export class RawValueFormatter implements ValueFormatter {
-  format(v: number) {
-    return v.toLocaleString()
-  }
-}
-
-export class TimeFormatter implements ValueFormatter {
-  private multiplier: number
-
-  constructor(unit: 'nanoseconds' | 'microseconds' | 'milliseconds' | 'seconds') {
-    if (unit === 'nanoseconds') this.multiplier = 1e-9
-    else if (unit === 'microseconds') this.multiplier = 1e-6
-    else if (unit === 'milliseconds') this.multiplier = 1e-3
-    else this.multiplier = 1
-  }
-
-  format(v: number) {
-    const s = v * this.multiplier
-
-    if (s / 60 >= 1) return `${(s / 60).toFixed(2)}min`
-    if (s / 1 >= 1) return `${s.toFixed(2)}s`
-    if (s / 1e-3 >= 1) return `${(s / 1e-3).toFixed(2)}ms`
-    if (s / 1e-6 >= 1) return `${(s / 1e-6).toFixed(2)}µs`
-    else return `${(s / 1e-9).toFixed(2)}ms`
-  }
-}
-
-export class ByteFormatter implements ValueFormatter {
-  format(v: number) {
-    if (v < 1024) return `${v.toFixed(2)} B`
-    v /= 1024
-    if (v < 1024) return `${v.toFixed(2)} KB`
-    v /= 1024
-    if (v < 1024) return `${v.toFixed(2)} MB`
-    v /= 1024
-    return `${v.toFixed(2)} GB`
-  }
-}
 
 export class Profile {
   protected name: string = ''
